+++ cf/coordinatereference.py
@@ -84,7 +84,7 @@ class CoordinateReference(cfdm.CoordinateReference):
     """
 
     def __new__(cls, *args, **kwargs):
-        """"""
+        """TODO."""
         instance = super().__new__(cls)
         instance._CoordinateConversion = CoordinateConversion
         instance._Datum = Datum