+++ cf/data/data.py
@@ -2040,7 +2040,7 @@ class Data(Container, cfdm.Data):
             self[indices] = u
 
     def _flag_partitions_for_processing(self, parallelise=True):
-        """"""
+        """TODO."""
         if mpi_on and parallelise:
             # Add a flag `_process_partition` to each partition defining
             # whether this partition will be processed on this process
@@ -7344,7 +7344,7 @@ class Data(Container, cfdm.Data):
         del self._custom["_all_axes"]
 
     def _flip(self, *flip):
-        """"""
+        """TODO."""
         if flip:
             self._custom["flip"] = flip[0]
         else:
@@ -12748,7 +12748,7 @@ class Data(Container, cfdm.Data):
         return d
 
     def HDF_chunks(self, *chunks):
-        """"""
+        """TODO."""
         _HDF_chunks = self._HDF_chunks
 
         if _HDF_chunks is None: