+++ cf/dimensioncoordinate.py
@@ -58,7 +58,7 @@ class DimensionCoordinate(
     """
 
     def __new__(cls, *args, **kwargs):
-        """"""
+        """TODO."""
         instance = super().__new__(cls)
         instance._Bounds = Bounds
         return instance