+++ cf/domainancillary.py
@@ -30,7 +30,7 @@ class DomainAncillary(mixin.PropertiesDataBounds, cfdm.DomainAncillary):
     """
 
     def __new__(cls, *args, **kwargs):
-        """"""
+        """TODO."""
         instance = super().__new__(cls)
         instance._Bounds = Bounds
         return instance