+++ cf/test/test_pp.py
@@ -18,7 +18,7 @@ tmpfiles = [
 
 
 def _remove_tmpfiles():
-    """"""
+    """TODO."""
     for f in tmpfiles:
         try:
             os.remove(f)